// web/src/app/shared/api-client/model/exercise-input-arrange-text-options-dto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ArrangeTextOption } from './arrange-text-option';
import { ExerciseInputDto } from './exercise-input-dto';


export interface ExerciseInputArrangeTextOptionsDto extends ExerciseInputDto { 
    value?: string;
    options: Array<ArrangeTextOption>;
}