// web/src/app/shared/api-client/model/exercise-input-pair-options-dto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { PairOptionsInputOptionDto } from './pair-options-input-option-dto';
import { ExerciseInputDto } from './exercise-input-dto';
import { PairOptionsInputFeedback } from './pair-options-input-feedback';


export interface ExerciseInputPairOptionsDto extends ExerciseInputDto { 
    leftOptionType: string;
    rightOptionType: string;
    leftOptions: Array<PairOptionsInputOptionDto>;
    rightOptions: Array<PairOptionsInputOptionDto>;
    feedback?: PairOptionsInputFeedback;
}