// web/src/app/shared/api-client/model/exercise-input-record-video-dto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ExerciseInputDto } from './exercise-input-dto';


export interface ExerciseInputRecordVideoDto extends ExerciseInputDto { 
    url: string;
}