// web/src/app/shared/api-client/model/exercise-input-select-placeholders-dto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { SelectPlaceholdersInputOption } from './select-placeholders-input-option';
import { ExerciseInputDto } from './exercise-input-dto';
import { SelectPlaceholdersInputPart } from './select-placeholders-input-part';


export interface ExerciseInputSelectPlaceholdersDto extends ExerciseInputDto { 
    parts: Array<SelectPlaceholdersInputPart>;
    options: Array<SelectPlaceholdersInputOption>;
}