// web/src/app/shared/api-client/model/exercise-input-write-placeholders-dto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { WritePlaceholdersFeedbackDto } from './write-placeholders-feedback-dto';
import { ExerciseInputDto } from './exercise-input-dto';
import { WritePlaceholdersPartDto } from './write-placeholders-part-dto';


export interface ExerciseInputWritePlaceholdersDto extends ExerciseInputDto { 
    parts?: Array<WritePlaceholdersPartDto>;
    feedback?: WritePlaceholdersFeedbackDto;
}